refactoring: tflite/cc/kernels -> tflite/kernels
