+++ examples/lce_minimal.cc
@@ -1,6 +1,6 @@
 #include <cstdio>
 
-#include "larq_compute_engine/tflite/cc/kernels/lce_ops_register.h"
+#include "larq_compute_engine/tflite/kernels/lce_ops_register.h"
 #include "tensorflow/lite/interpreter.h"
 #include "tensorflow/lite/kernels/register.h"
 #include "tensorflow/lite/model.h"
