+++ larq_compute_engine/tflite/benchmark/BUILD
@@ -23,7 +23,7 @@ cc_binary(
         "//conditions:default": [],
     }),
     deps = [
-        "//larq_compute_engine/tflite/cc/kernels:lce_op_kernels",
+        "//larq_compute_engine/tflite/kernels:lce_op_kernels",
         "@org_tensorflow//tensorflow/lite/tools/benchmark:benchmark_tflite_model_lib",
         "@org_tensorflow//tensorflow/lite/tools/benchmark:logging",
     ],
@@ -43,7 +43,7 @@ cc_binary(
         "//conditions:default": [],
     }),
     deps = [
-        "//larq_compute_engine/tflite/cc/kernels:lce_op_kernels",
+        "//larq_compute_engine/tflite/kernels:lce_op_kernels",
         "@org_tensorflow//tensorflow/lite:framework",
         "@org_tensorflow//tensorflow/lite/kernels:builtin_ops",
         "@googlebenchmark//:benchmark",
