+++ larq_compute_engine/tflite/benchmark/lce_benchmark_all.cc
@@ -1,6 +1,6 @@
 #include <benchmark/benchmark.h>
 
-#include "larq_compute_engine/tflite/cc/kernels/lce_ops_register.h"
+#include "larq_compute_engine/tflite/kernels/lce_ops_register.h"
 #include "tensorflow/lite/interpreter.h"
 #include "tensorflow/lite/kernels/register.h"
 #include "tensorflow/lite/model.h"
