+++ larq_compute_engine/tflite/kernels/bgemm_trmul_params.h
