+++ larq_compute_engine/tflite/kernels/lce_register.cc
@@ -6,7 +6,7 @@
 // but this way we only have to add them once.
 //
 
-#include "larq_compute_engine/tflite/cc/kernels/lce_ops_register.h"
+#include "larq_compute_engine/tflite/kernels/lce_ops_register.h"
 #include "tensorflow/lite/kernels/builtin_op_kernels.h"
 #include "tensorflow/lite/kernels/register.h"
