+++ larq_compute_engine/tflite/tests/BUILD
@@ -8,7 +8,7 @@ cc_test(
     size = "small",
     srcs = ["bconv2d_test.cc"],
     deps = [
-        "//larq_compute_engine/tflite/cc/kernels:lce_ops",
+        "//larq_compute_engine/tflite/kernels:lce_ops",
         "@org_tensorflow//tensorflow/lite/kernels:test_main",
         "@org_tensorflow//tensorflow/lite/kernels:test_util",
         "@org_tensorflow//tensorflow/lite:framework",
